backend/auto-news/requirements.txt
openai==1.54.0
psycopg2-binary==2.9.9
